requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
pandas
openpyxl
# Google Sheets & Auth
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
oauthlib==3.2.2
